+++ odoo/local-src/db2_import/models/__init__.py
@@ -1,1 +1,3 @@
 from . import db2_importer
+from . import inventory
+from . import queue_job